moved file, and simplified the loot of a golem
